升级1.1

+++ BaoKanIOS/BaoKanIOS/Classes/Module/News/Controller/JFNewsViewController.swift
@@ -370,11 +370,11 @@ extension JFNewsViewController: UIScrollViewDelegate {
         let scaleLeft = 1 - scaleRight
         
         let labelLeft = topScrollView.subviews[leftIndex] as! JFTopLabel
-        labelLeft.scale = Float(scaleLeft)
+        labelLeft.scale = CGFloat(scaleLeft)
         
         if rightIndex < topScrollView.subviews.count {
             let labelRight = topScrollView.subviews[rightIndex] as! JFTopLabel
-            labelRight.scale = Float(scaleRight)
+            labelRight.scale = CGFloat(scaleRight)
         }
         
         var index = Int(value)
