+++ BaoKanIOS/BaoKanIOS/Classes/Module/Photo/Controller/JFPhotoViewController.swift
@@ -192,11 +192,11 @@ extension JFPhotoViewController: UIScrollViewDelegate {
         let scaleLeft = 1 - scaleRight
         
         let labelLeft = topScrollView.subviews[leftIndex] as! JFTopLabel
-        labelLeft.scale = Float(scaleLeft)
+        labelLeft.scale = CGFloat(scaleLeft)
         
         if rightIndex < topScrollView.subviews.count {
             let labelRight = topScrollView.subviews[rightIndex] as! JFTopLabel
-            labelRight.scale = Float(scaleRight)
+            labelRight.scale = CGFloat(scaleRight)
         }
         
         // 计算子控制器数组角标
